--- test.c
@@ -1,10 +1,0 @@
-e
-e
-e
-e
-e
-e
-e
-e
-e
-e
